refactor(uploads): extract active transfer owner
