+++ scripts/ci/module-impact.json
@@ -90,7 +90,11 @@
       "fallbackCapability": "main_runtime"
     },
     "upload_repository": {
-      "ownerPaths": ["src/main/uploads/repository.ts"],
+      "ownerPaths": [
+        "src/main/uploads/repository.ts",
+        "src/main/uploads/active-transfer-owner.ts",
+        "src/main/uploads/storage-helpers.ts"
+      ],
       "interfacePaths": ["src/main/uploads/repository.ts"],
       "consumerModules": ["session_persistence"],
       "testFiles": {
@@ -99,7 +103,11 @@
           "src/main/uploads/repository.characterization.test.ts"
         ],
         "contract": ["src/main/uploads/ipc.test.ts", "src/main/uploads/command-owner.test.ts"],
-        "consumer": ["src/main/session-persistence/deletion-integration.test.ts"]
+        "consumer": [
+          "src/main/session-persistence/deletion-integration.test.ts",
+          "src/main/acp/file-reference-resolver.test.ts",
+          "src/main/acp/prompt-content-owner.test.ts"
+        ]
       },
       "capabilityOverlays": ["windows_sensitive"],
       "fallbackCapability": "main_runtime"
